fixed visualizer file management and intersection checks

+++ src/solutionManagement.py
@@ -12,6 +12,7 @@ def loadSolutions(foldername):
     fn = foldername.name
     summaryName = foldername.parent.parent/"solution_summaries"/(str(fn)+".zip")
     pickelName = foldername.parent.parent/"solution_summaries"/(str(fn)+".pkl")
+    addedSolutions = False
     if summaryName.exists():
         i = 0
         names = pickle.load(open(pickelName, "rb"))
@@ -22,7 +23,7 @@ def loadSolutions(foldername):
             elif len(sol.steiner_points_x) < len(best[i][0].steiner_points_x):
                 assert sol.instance_uid == best[i][0].instance_uid
                 best[i] = [sol,names[i]]
-                logging.info(str(names[i])+" is better at "+str(sol.instance_uid) +" with solution size "+str(len(sol.steiner_points_x)))
+                #logging.info(str(names[i])+" is better at "+str(sol.instance_uid) +" with solution size "+str(len(sol.steiner_points_x)))
             i += 1
 
     #first build the list
@@ -32,8 +33,12 @@ def loadSolutions(foldername):
             logging.info("reading "+str(solname))
             for sol in ZipSolutionIterator(solname):
                 if len(best) == i:
+                    if addedSolutions == False:
+                        addedSolutions = True
                     best.append([sol,solname])
                 elif len(sol.steiner_points_x) < len(best[i][0].steiner_points_x):
+                    if addedSolutions == False:
+                        addedSolutions = True
                     assert sol.instance_uid == best[i][0].instance_uid
                     best[i] = [sol,solname]
                     logging.info(str(solname)+" is better at "+str(sol.instance_uid) +" with solution size "+str(len(sol.steiner_points_x)))
@@ -43,22 +48,23 @@ def loadSolutions(foldername):
         logging.error("No matching data found for folder "+str(foldername))
         return []
 
-    #now rebuild summary
-    if summaryName.exists():
-        summaryName.unlink()
+    if addedSolutions:
+        #now rebuild summary
+        if summaryName.exists():
+            summaryName.unlink()
 
-    if pickelName.exists():
-        pickelName.unlink()
-    #Write the solutions to a new zip file
-    toWriteNames = []
-    with ZipWriter(summaryName) as zw:
-        logging.info("writting sol summary at "+str(summaryName))
-        for solution,name in best:
-            zw.add_solution(solution)
-            toWriteNames.append(name)
-    with open(pickelName, "wb") as f:
-        logging.info("writting name summary at "+str(pickelName))
-        pickle.dump(toWriteNames, f)
+        if pickelName.exists():
+            pickelName.unlink()
+        #Write the solutions to a new zip file
+        toWriteNames = []
+        with ZipWriter(summaryName) as zw:
+            logging.info("writting sol summary at "+str(summaryName))
+            for solution,name in best:
+                zw.add_solution(solution)
+                toWriteNames.append(name)
+        with open(pickelName, "wb") as f:
+            logging.info("writting name summary at "+str(pickelName))
+            pickle.dump(toWriteNames, f)
 
     return best
